Revert "Merge branch 'main' into Duc"
This reverts commit 73e33c3c9e038631c2ff25d39e4c291375eb8273, reversing
changes made to 1ee0b62f13ae92606fbc9b38f9e4976ca6597b25.

+++ database/migrations/2024_06_05_073823_create_vouchers_table.php
@@ -12,9 +12,9 @@ class CreateVouchersTable extends Migration
             $table->bigIncrements('id');
             $table->string('code')->unique();
             $table->string('discount_type');
-            $table->integer('discount_value');
+            $table->decimal('discount_value', 8, 2);
             $table->date('expiry_date');
-            $table->integer('min_purchase_amount')->nullable();
+            $table->decimal('min_purchase_amount', 8, 2)->nullable();
             $table->unsignedBigInteger('category_id')->nullable();
             $table->integer('max_usage')->nullable();
             $table->integer('used_count')->default(0);
